+++ src/buttons/panel_hub_gluecksrad.js
@@ -1,13 +1,9 @@
+const { createGluecksradHubMessage } = require("../utils/panelHub");
+
 module.exports = {
   customId: "panel_hub_gluecksrad",
 
   async execute(interaction) {
-    return interaction.reply({
-      content:
-        "🎡 **Glücksrad Panel**\n\n" +
-        "Zum Erstellen nutze:\n" +
-        "`/gluecksradpanel`",
-      flags: 64
-    });
+    return interaction.update(createGluecksradHubMessage());
   }
 };