+++ src/buttons/panel_hub_home.js
@@ -1,0 +1,9 @@
+const { createPanelHubMessage } = require("../utils/panelHub");
+
+module.exports = {
+  customId: "panel_hub_home",
+
+  async execute(interaction) {
+    return interaction.update(createPanelHubMessage());
+  }
+};