+++ src/buttons/panel_hub_music.js
@@ -1,13 +1,9 @@
+const { createMusicHubMessage } = require("../utils/panelHub");
+
 module.exports = {
   customId: "panel_hub_music",
 
   async execute(interaction) {
-    return interaction.reply({
-      content:
-        "🎵 **Music Panel**\n\n" +
-        "Zum Erstellen oder Aktualisieren nutze:\n" +
-        "`/musicpanel`",
-      flags: 64
-    });
+    return interaction.update(createMusicHubMessage());
   }
 };